+++ Win32/Hooks/Display.cs
@@ -2,12 +2,12 @@
 {
     public class Display
     {
-        public double X { get; set; }
-        public double Y { get; set; }
-        public double Width { get; set; }
-        public double Height { get; set; }
+        public int X { get; set; }
+        public int Y { get; set; }
+        public int Width { get; set; }
+        public int Height { get; set; }
 
-        public Display(double x, double y, double width, double height)
+        public Display(int x, int y, int width, int height)
         {
             X = x;
             Y = y;
